proto4z reserve field.  http on block integrity check.

+++ example/udpTest/main.cpp
@@ -163,7 +163,7 @@ int main(int argc, char* argv[])
                 std::placeholders::_3,
                 std::placeholders::_4,
                 pic));
-            pic->sock->doSendTo((char *)g_fillString.c_str(), g_fillString.length(), ip.c_str(), port);
+            pic->sock->doSendTo((char *)g_fillString.c_str(), (unsigned int)g_fillString.length(), ip.c_str(), port);
         }
 
     }
